Fix XCode simulator build for XCode 6.1
When building with XCode 6.1, the Simulator build will fail the
configure step, due to gcc checks seemingly either not pulling in
a needed library that defines an init function, or that it doesn't
explicitly define the needed init function.

To correct this, setting the minimum IOS version to 6.0 fixes this.

Of course, this limits users of the curl library to a minimum of
version 6.0.